+++ src/main/scala/io.mth.pirate.internal/ParseState.scala
@@ -1,0 +1,5 @@
+package io.mth.pirate.internal
+
+sealed trait ParseState
+case object SkipOpts extends ParseState
+case object AllowOpts extends ParseState
