+++ src/main/scala/io.mth.pirate/Metadata.scala
@@ -1,0 +1,3 @@
+package io.mth.pirate
+
+case class Metadata(description: Option[String], visible: Boolean)
