+++ src/main/scala/io.mth.pirate/ParseError.scala
@@ -1,0 +1,5 @@
+package io.mth.pirate
+
+sealed trait ParseError
+case object ParseErrorNoMessage extends ParseError
+case class ParseErrorMessage(s: String) extends ParseError
