+++ src/main/scala/io.mth.pirate/Prefs.scala
@@ -1,0 +1,4 @@
+package io.mth.pirate
+
+sealed trait Prefs
+case object NullPrefs extends Prefs
