+++ src/main/scala/io.mth.pirate/Runners.scala
@@ -1,18 +1,19 @@
 package io.mth.pirate
 
 import scalaz._, Scalaz._, effect.IO
+import io.mth.pirate.internal._
 
 trait Runners {
-  def run[A, B](args: List[String], command: Command[A], f: A => B): PError \/ B  =
+  def run[A, B](args: List[String], command: Command[A], f: A => B): ParseError \/ B  =
     Interpretter.run(command.parse, args).map(f)
 
   def runOrFail[A](args: List[String], command: Command[A], f: A => IO[Unit]): IO[Unit]  =
     run[A, IO[Unit]](args, command, f) match {
       case -\/(e) => e match {
-        case PErrorNoMessage => IO {
+        case ParseErrorNoMessage => IO {
           Console.err.print(Usage.print(command))
         }
-        case PErrorMessage(s) => IO {
+        case ParseErrorMessage(s) => IO {
           Console.err.println(s)
           Console.err.print(Usage.print(command))
         }
